Ai_Agent-main/requirements.txt
flask>=2.0.0
openai>=0.27.0
mysql-connector-python>=8.0.0
python-dotenv>=0.19.0
tiktoken>=0.4.0


# pip install -r requirements.txt
